% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccEstimate.R
\name{ccEstimate}
\alias{ccEstimate}
\title{Calculate cell confluency.}
\usage{
ccEstimate(
  infile,
  reduced_size = 0.8,
  sharpen_filter_size = 15,
  contrast = 2,
  brush_size = 5,
  shape = "disc",
  random_image_size = 50,
  n_random_images = 100
)
}
\arguments{
\item{infile}{Input file (JPG, PNG or TIFF).}

\item{reduced_size}{Use a sub-image with dimensions reduced_size * width: reduced_size * height. Default = 0.8.}

\item{sharpen_filter_size}{Size of the sharpen filter. Creates a squared matrix with size sharpen_filter_size: values are -1 on diagonals and center = sharpen_filter_size * 2 -2. Default = 15.}

\item{contrast}{Contrast value. Default = 2.}

\item{brush_size}{Brush size. Default = 5.}

\item{shape}{Brush shape. Can be box, disc, diamond, Gaussian or line. Default = "disc".}

\item{random_image_size}{Size of random sub-images to be generated. Default = 50.}

\item{n_random_images}{Number of random images to be generated. Default = 100.}
}
\value{
The confluency level of the image in the input file.
}
\description{
Calculate cell confluency.
}
\examples{
example_low_confluency = system.file("extdata", "low_confluency.jpg", package = "ccEstimate")
example_high_confluency = system.file("extdata", "high_confluency.jpg", package = "ccEstimate")
ccEstimate(example_low_confluency)
ccEstimate(example_high_confluency)
}
