% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccEstimate.R
\name{run_confluency_estimation}
\alias{run_confluency_estimation}
\title{Estimate when confluency will reach a predetermined threshold.}
\usage{
run_confluency_estimation(
  infolder = "",
  run_example = FALSE,
  confluence_target = 0.8,
  plot_confluence = TRUE
)
}
\arguments{
\item{infolder}{Input folder where all the images to analyze are stored. Deafult = "".}

\item{run_example}{Run example data. Default = FALSE.}

\item{confluence_target}{Confluence at which differentiation should start. Values between 0 and 1. Default = 0.8.}

\item{plot_confluence}{Plot confluence values at each time point. Default = TRUE}
}
\value{
A data frame with the confluency estimation for each sample and flask analyzed
}
\description{
Estimate when confluency will reach a predetermined threshold.
}
\examples{
run_confluency_estimation(run_example = TRUE)
}
